backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlparse
pandas
numpy
scikit-learn